Added Dale-Chall and Spache scoring
Initial tests indicate these are working well, but they need some more
formal unit tests. They assumes that the plural of a known easy word is
still easy, which I am not certain is correct. Other tools also include "-
ly", "-d", "-r", "-ing" etc as easy words - need to confirm if that is
correct (and if it is, to modify the calculator).